Moved the TabLayout to the fragment xml to support anchoring of TabLayout when scrolling
