+++ WooCommerce/src/main/res/layout/dashboard_stats.xml
@@ -6,24 +6,10 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.google.android.material.tabs.TabLayout
-        android:id="@+id/tab_layout"
-        style="@style/Woo.TabLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"/>
-
-    <FrameLayout
-        android:id="@+id/tab_layout_divider"
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:background="@color/list_divider"
-        app:srcCompat="@drawable/list_divider"/>
-
     <LinearLayout
         android:id="@+id/label_layout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
         android:baselineAligned="false"
         android:gravity="center_horizontal"
         android:orientation="horizontal">
